Update README.md: Remove disguises
I don't believe we should have disguises in this plugin, since that would make it CB-version-dependent. It would probably be best that server owners simply go get a different disguise plugin and grant those permissions and such, so then they don't have to wait for this plugin (and LibDisguises [or whatever we'd be using]) to update if they want to update their servers.