Added customizable messages:
Messages.yml now contains a bunch of nodes the server owners can change
to suit their language or, if they just don't like our colours.

I suggest that all developers use this for their messages. I have
updated @pookeythekid 's mob cannon so that it uses it.

The standard format for the gadget's names should be
- Gadgets:
   - <Gadget Name>:
      -name: "Gadget Name"