Paintball Gadget:
I'm really tired and I don't know if this is the best way of
implementing PainballGuns into the plugin but, it seems to work.

I originally wanted to use the Player#sendBlockChange but, didn't know
whether the block changes should've been per player or server wider so,
I just went with server wide changes.