Update messages.yml
Added "description" fields for the gadgets. These will be added to the item's lore when in the gadget selector GUI and removed when/if the player receives the gadget (Maybe copy without the lore?)

I will also add variables such as "{COST}", "{BALANCE}" to display cost of gadget (If server owner wants that) and the balance of player.
If you want me to add any other variable similar to the above, either comment with the variable (e.g. {VARIABLE}) and what it should represent (e.g. Player's health).