Changed the way disguises are implemented. Instead of being dependant on reflection utils, disguises can now have direct access to NMS code. To implement, just do the following:
1. Create folder in "ovh.tgrhavoc.mvpgadgets.disguisegadget.nms" that is
the name of the version you want to implment (e.g. v1_8_R3 for 1.8.8).
2. Add "Disguise.java" (Case-sensitive)
3. Implement the methods, using the NMS code for you perticular version.
4. Win?

I think this way just gives the disguise gadget a lot more flexability
in the sense that implementing the gadget for different server versions
is a lot more easy and doesn't rely on them having the same structure
and class names.