Let's get started... Again
Set the default values to "guiGadgetWorlds" to an empty array, just to allow users to use the plugin out-of-the-box. I've also added a new config option to allow the users to set whether the players can drop the gadget items or not.

Because of the added config option I've also added a new message to tell the players they cannot drop items.