1.8 Release (#40)
* Let's get started... Again

Set the default values to "guiGadgetWorlds" to an empty array, just to allow users to use the plugin out-of-the-box. I've also added a new config option to allow the users to set whether the players can drop the gadget items or not.

Because of the added config option I've also added a new message to tell the players they cannot drop items.

* Gadget registration

Re-done how the gadgets are registered on the server. The main gadget class (Gadget.java) checks to see if it's already called the Gadget#registerEvent method and shouldn't call it multiple times.

* 1.8 Stable

Plugin should now be stable on 1.8.x servers (If not, then at least 1.8.4 and above).

* 1.8 Stable

Plugin should now be stable on 1.8.x servers (If not, then at least 1.8.4 and above).