Minecraft 1.9.4
Added disguise and paintball support for v1_9_R2 NMS (1.9.4).

Apparently the disguise gadget was using the deprecated disguise class for the disguise class names.. So I removed it..

The horse gadget managed to screw up the server when you dismount a horse in mid air... I fixed it but at a price. Players can now spam-spawn horses.