Update to 1.3.0 (#41)
* Let's get started... Again

Set the default values to "guiGadgetWorlds" to an empty array, just to allow users to use the plugin out-of-the-box. I've also added a new config option to allow the users to set whether the players can drop the gadget items or not.

Because of the added config option I've also added a new message to tell the players they cannot drop items.

* Gadget registration

Re-done how the gadgets are registered on the server. The main gadget class (Gadget.java) checks to see if it's already called the Gadget#registerEvent method and shouldn't call it multiple times.

* 1.8 Stable

Plugin should now be stable on 1.8.x servers (If not, then at least 1.8.4 and above).

* 1.8 Stable

Plugin should now be stable on 1.8.x servers (If not, then at least 1.8.4 and above).

* Minecraft 1.9.2

Disguises and Paintball should now work on 1.9 & 1.9.2 (maybe 1.9.1).

The paintball gadget now uses an abstract hanlder to send the updates to the blocks. You can get the handler by calling PaintFactory.getPaintHandler( String ).

* Updated sound effects

Apparently 1.9 and above have diferent sound names. Explode is now called Entity_generic_exploded.

* Minecraft 1.9.4

Added disguise and paintball support for v1_9_R2 NMS (1.9.4).

Apparently the disguise gadget was using the deprecated disguise class for the disguise class names.. So I removed it..

The horse gadget managed to screw up the server when you dismount a horse in mid air... I fixed it but at a price. Players can now spam-spawn horses.

* Minecraft 1.10 & 1.10.2

We're now up-to-date with all the minecraft servers.. I think..

* Updated list of entities in config

* Removed System.out from EntityDisguises